Move Image Gallery to a separate component
